Move data proxy
